ProjectileHandler
I fixed my git finally, had to reinstall and wipe all my stuff off disk
related to project, not sure what caused the problem. This has both the
base projectile and its handler ready, if you want when we have a basic
usable weapon I will put in an example of each different projectile type
so that it is easier for you to implement your own weapons.