Various Bug Fixes and Cowtapult, still missing texture for that
By the way Jay, in NPC I tried to set the onabduct and ondeath method
with the scripts but whenever I included scripthandler to get its
pointer it crashed the entire NPC class. I commented out the include and
appropriate method calls in onabduct/ondeath if you can fix the crashing
with the include statement.